+++ tests/conftest.py
@@ -5,23 +5,31 @@ import shutil
 import tempfile
 from importlib.metadata import version
 from typing import Iterator
+from types import MappingProxyType
 from unittest.mock import patch
 
 import pytest
 from openghg.retrieve import search
-from openghg.standardise import standardise_surface, standardise_bc, standardise_flux, standardise_footprint, standardise_column
+from openghg.standardise import (
+    standardise_surface,
+    standardise_bc,
+    standardise_flux,
+    standardise_footprint,
+    standardise_column,
+)
 from openghg.types import ObjectStoreError
 import xarray as xr
 import zarr
 
-
 _raw_data_path = Path(".").resolve() / "tests/data/"
+
 
 @pytest.fixture(scope="session", autouse=True)
 def user_data_path():
     _user_data_path = Path(tempfile.gettempdir()) / f"{getpass.getuser()}_openghg_inversions_test_data"
     _user_data_path.mkdir(exist_ok=True)
     return _user_data_path
+
 
 @pytest.fixture(scope="session")
 def openghg_version():
@@ -96,10 +104,12 @@ def country_ds_eastasia(raw_data_path):
     ds = xr.load_dataset(raw_data_path / "country_EASTASIA.nc")
     yield ds
 
+
 @pytest.fixture
 def southamerica_country_file(raw_data_path):
     """Provides path to the SOUTHAMERICA countryfile"""
-    return raw_data_path /"satellite"/"country"/ "country_SOUTHAMERICA.nc"
+    return raw_data_path / "satellite" / "country" / "country_SOUTHAMERICA.nc"
+
 
 @pytest.fixture(scope="session", autouse=True)
 def session_config_mocker(user_data_path) -> Iterator[None]:
@@ -146,23 +156,34 @@ test_data_list.append(TestData(standardise_surface, mhd_obs_metadata, mhd_obs_da
 ## Satellite Column data
 satellite_gosat_obs_metadata = {
     "source_format": "openghg",
-    "satellite":"gosat",
+    "satellite": "gosat",
     "network": "gosat",
-    "domain":"southamerica",
-    "instrument":"tanso-fts",
-    "species":"ch4",
+    "domain": "southamerica",
+    "instrument": "tanso-fts",
+    "species": "ch4",
 }
-satellite_gosat_obs_data_path = _raw_data_path / "satellite"/"column"/"gosat-fts_gosat_20160101_ch4-column.nc"
-test_data_list.append(TestData(standardise_column, satellite_gosat_obs_metadata, satellite_gosat_obs_data_path, "column"))
+satellite_gosat_obs_data_path = (
+    _raw_data_path / "satellite" / "column" / "gosat-fts_gosat_20160101_ch4-column.nc"
+)
+test_data_list.append(
+    TestData(standardise_column, satellite_gosat_obs_metadata, satellite_gosat_obs_data_path, "column")
+)
 
 ## BC data
 bc_metadata = {"species": "ch4", "bc_input": "cams", "domain": "europe", "store": "inversions_tests"}
 bc_data_path = _raw_data_path / "bc_ch4_europe_cams_2019-01-01_2019-12-31_data.nc"
 test_data_list.append(TestData(standardise_bc, bc_metadata, bc_data_path, "boundary_conditions"))
 
-satellite_bc_metadata = {"species": "ch4", "bc_input": "cams", "domain": "southamerica", "store": "inversions_tests"}
-satellite_bc_data_path = _raw_data_path / "satellite"/ "bc" /"ch4_SOUTHAMERICA_201601_CAMS-inversion.nc"
-test_data_list.append(TestData(standardise_bc, satellite_bc_metadata, satellite_bc_data_path, "boundary_conditions"))
+satellite_bc_metadata = {
+    "species": "ch4",
+    "bc_input": "cams",
+    "domain": "southamerica",
+    "store": "inversions_tests",
+}
+satellite_bc_data_path = _raw_data_path / "satellite" / "bc" / "ch4_SOUTHAMERICA_201601_CAMS-inversion.nc"
+test_data_list.append(
+    TestData(standardise_bc, satellite_bc_metadata, satellite_bc_data_path, "boundary_conditions")
+)
 
 ## Footprint data
 tac_footprints_metadata = {
@@ -173,7 +194,9 @@ tac_footprints_metadata = {
     # "metmodel": "ukv",
 }
 tac_footprints_data_path = _raw_data_path / "footprints_tac_europe_name_185m_2019-01-01_2019-01-07_data.nc"
-test_data_list.append(TestData(standardise_footprint, tac_footprints_metadata, tac_footprints_data_path, "footprints"))
+test_data_list.append(
+    TestData(standardise_footprint, tac_footprints_metadata, tac_footprints_data_path, "footprints")
+)
 
 mhd_footprints_metadata = {
     "site": "mhd",
@@ -183,8 +206,10 @@ mhd_footprints_metadata = {
     "source_format": "paris",
     # "metmodel": "ukv",
 }
-mhd_footprints_data_path = _raw_data_path/"footprints_mhd_europe_name_10m_2019-01-01_2019-01-07_data.nc"
-test_data_list.append(TestData(standardise_footprint, mhd_footprints_metadata, mhd_footprints_data_path, "footprints"))
+mhd_footprints_data_path = _raw_data_path / "footprints_mhd_europe_name_10m_2019-01-01_2019-01-07_data.nc"
+test_data_list.append(
+    TestData(standardise_footprint, mhd_footprints_metadata, mhd_footprints_data_path, "footprints")
+)
 
 footprints_satellite_metadata = {
     "satellite": "GOSAT",
@@ -193,11 +218,14 @@ footprints_satellite_metadata = {
     "inlet": "column",
     "source_format": "acrg_org",
     "obs_region": "brazil",
-    "species": "ch4"
-
+    "species": "ch4",
 }
-footprints_satellite_data = _raw_data_path /"satellite"/ "footprints" / "GOSAT-BRAZIL-column_SOUTHAMERICA_201601.nc"
-test_data_list.append(TestData(standardise_footprint, footprints_satellite_metadata, footprints_satellite_data, "footprints"))
+footprints_satellite_data = (
+    _raw_data_path / "satellite" / "footprints" / "GOSAT-BRAZIL-column_SOUTHAMERICA_201601.nc"
+)
+test_data_list.append(
+    TestData(standardise_footprint, footprints_satellite_metadata, footprints_satellite_data, "footprints")
+)
 
 ## Flux data
 flux_metadata = {"species": "ch4", "source": "total-ukghg-edgar7", "domain": "europe"}
@@ -208,12 +236,16 @@ flux_dim_shuffle_metadata = {"species": "ch4", "source": "total-ukghg-edgar7-shu
 flux_dim_shuffled_data_path = (
     _raw_data_path / "flux_total_ch4_europe_edgar7_2019-01-01_2019-12-31_data_dim_shuffled.nc"
 )
-test_data_list.append(TestData(standardise_flux, flux_dim_shuffle_metadata, flux_dim_shuffled_data_path, "flux"))
+test_data_list.append(
+    TestData(standardise_flux, flux_dim_shuffle_metadata, flux_dim_shuffled_data_path, "flux")
+)
 
-flux_satellite_metadata = {"species":"ch4", "source":"SWAMPS",
-                           "domain":"southamerica"}
-flux_satellite_datapath = (_raw_data_path/"satellite"/ "flux"/"ch4_SOUTHAMERICA_2016_SWAMPS-v32-5_Saunois-Annual-Mean.nc")
+flux_satellite_metadata = {"species": "ch4", "source": "SWAMPS", "domain": "southamerica"}
+flux_satellite_datapath = (
+    _raw_data_path / "satellite" / "flux" / "ch4_SOUTHAMERICA_2016_SWAMPS-v32-5_Saunois-Annual-Mean.nc"
+)
 test_data_list.append(TestData(standardise_flux, flux_satellite_metadata, flux_satellite_datapath, "flux"))
+
 
 @pytest.fixture(scope="session", autouse=True)
 def session_object_store(session_config_mocker) -> None:
@@ -239,7 +271,6 @@ def session_object_store(session_config_mocker) -> None:
             add_data = True
         else:
             add_data = Counter([x.data_type for x in test_data_list]) != Counter(found_dtypes)
-
 
     # check if there are four pieces of data in the object store
     # if not, add the missing data
@@ -298,11 +329,12 @@ def tac_ch4_data_args():
     }
     return data_args
 
+
 @pytest.fixture(scope="module")
 def satellite_ch4_data_args():
     data_args = {
-        "species" : "ch4",
-        "sites": ['GOSAT-BRAZIL'], 
+        "species": "ch4",
+        "sites": ["GOSAT-BRAZIL"],
         "averaging_period": ["1H"],
         "start_date": "2016-01-01",
         "end_date": "2016-02-01",
@@ -322,6 +354,7 @@ def satellite_ch4_data_args():
         # "met_model": "ukv",
     }
     return data_args
+
 
 @pytest.fixture(scope="module")
 def mhd_and_tac_ch4_data_args():
@@ -344,3 +377,26 @@ def mhd_and_tac_ch4_data_args():
         "averaging_period": ["1h", "1h"],
     }
     return data_args
+
+
+@pytest.fixture(scope="module")
+def mhd_and_tac_fp_data(mhd_and_tac_ch4_data_args):
+    from openghg_inversions.basis import basis_functions_wrapper
+    from openghg_inversions.inversion_data.get_data import data_processing_surface_notracer
+
+    fp_all, *_ = data_processing_surface_notracer(**mhd_and_tac_ch4_data_args)
+
+    basis_args = {
+        "species": "ch4",
+        "domain": "EUROPE",
+        "start_date": "2019-01-01",
+        "emissions_name": ["total-ukghg-edgar7"],
+        "nbasis": 20,
+        "use_bc": True,
+        "basis_algorithm": "weighted",
+        "bc_basis_case": "NESW",
+    }
+
+    fp_data = basis_functions_wrapper(fp_all, **basis_args)
+
+    return MappingProxyType(fp_data)  # read-only